Steam vortices trail steam and make a cloud of steam when killed
This aims to make steam vortices into a monster that plays a little
differently, no longer just a popcorn monster that you brush off and
don't really care about. Their trailing clouds are temporary, but block
line of sight for a little bit, and once killed they deny vision of the
whole area for a few turns, which both have some tactical implications.