Add the #shout command
Most modern variants have this already so this is playing catch-up. This
command basically allows the player to print a string of their choice to
the livelog, which in 3.7 also places it in the game chronicle. It has
applications in IRC integration and roleplay (imagine screaming your
victory to the heavens after you slay your way through Thoth Amon's
horde and recover the Heart of Ahriman, for instance).

The unique thing about this implementation (I think) is that it has an
in-game effect: assuming your current form is capable of producing
sound, you do actually shout in-game, which will wake up monsters. This
does give Knights a free way to wake up rooms of sleeping monsters to
avoid alignment penalties, but they have had plenty of ways to do that
already, such as kicking the nearest wall, so it doesn't really affect
much. If you *cannot* speak, you shout mentally and it still gets
recorded in the livelog, but makes no noise.

Complicating the matter on the code side is that vanilla had already
used up all the livelog flags in the 0xffff space, so this needed to add
a new one at 0x10000 and push the boundaries in the code just a little
bit. The sysconf must enable this 0x10000 flag in LIVELOG or else it
won't show up on anything.

Initially I was going to limit the amount of recorded shouts to 20, but
then decided players have generally not abused them much in other
variants on Hardfought, so for now we'll go with it being unlimited.

Not sure if shouts will show up in the dumplog; I haven't tested.