XL10+ Rangers automatically see ammo enchantments
This is a minor role benefit for Rangers, which will hopefully make them
more able to use ammo they find.

Note that they need to have Basic skill with the ammo in order to see
the enchantment; this was a last-minute addition but I decided that if
your Ranger has never touched a crossbow, they probably shouldn't be
experienced enough with crossbow bolts to see their enchantment.

Code-wise, I didn't want to stick this as a side-effect in xname, but
then realized other things like Priest auto-beatitude-identification
also was done there, so it already had side effects anyway.