Make yellow and black light explosions directionless
Primarily to make them a little more interesting tactically. Now, for
instance, you can apply a blinding or confusing status effect to a group
of monsters, particularly if you can confuse or mislead the light into
exploding when it's not adjacent to you. Or if a yellow light chases you
down at the same time as some other monsters, those other monsters might
be blundering around as well and give you an opportunity to escape. A
secondary reason is realism: they're a ball of light that has barely
enough awareness to follow a target until adjacent to it and then blow
up, and they're not really capable of directing it a certain way.

Pet lights will *not* consider whether you are going to be affected by
their explosion, since they are after all mindless. (The pet lights
conjured by the confused scroll of light are cancelled and incapable
of exploding.)

The new function resists_light_halu is intended as a centralization of
the inconsistent checks around whether a given monster was subject to a
black light's attack.

Also contains some refactoring of explmm() since one if statement was
rendered unreachable by these changes.