Make the hunger property halve nutrition from food
This is aimed at giving a unique fringe benefit to the ring of hunger,
which was formerly an unwanted item in practically every circumstance.
It was only really useful for getting unsatiated while eating a lot of
nonperishable food, and even then other rings such as conflict could do
the same thing. It's now... well, still primarily mainly useful for
eating a lot of food, but you won't *get* satiated as fast, which makes
it useful for, say, eating corpses without letting them spoil.

This change does make it nastier when cursed. If it turns out to be a
death sentence to put a cursed ring on, I might have to revisit this.

Drinks are unaffected mainly because they don't go through lesshungry().
Basically, this only works on things that you could choke on. (Note that
this doesn't *prevent* choking - it'll just take more food to get to
that point.)