Make fuzzer logging a compile time switch and have it off by default
This has been bothering me for a while: games on Hardfought were eating
up CPU cycles and RAM for the sake of writing to a fuzzer log that was
never intended to be viewed in anything other than gdb. So this hides
the entire system behind #define FUZZER_LOG: without that enabled at
compile time, the game has no concept of a fuzzer log at all. I decided
to go the route of wrapping every call in that rather than just the guts
of the logging, because otherwise there would still be wasted effort
preparing string buffers and making function calls that would never get
used.

FUZZER_LOG is now turned on only in the linux-xnh-debug hints file. If
you use a different hints file and want fuzzer logging, you must enable
it there.

As a result of these changes, I had to turn remove_monster and
place_worm_seg into actual functions rather than macros, because you
can't have an ifdef inside a multiline macro. This is probably for the
best anyway, and I'm sure the CPU overhead incurred by having to push
stack frames for these functions is dwarfed by the savings on not having
to do the fuzzer log stuff.