Identify wands of probing and secret door detection when engraved with
This is from a suggestion by Red kangaroo. There's no compelling reason
for the large pools of wands that are ambiguous to identify should stay
that large, and these two came with ready-made messages. (I suspect the
only thing standing in the way of undead turning, opening, and locking
identifying when engraved with is that no one has thought up a good
engraving-effect message for them yet.)