Force "Is this ok?" prompt and show polyinit form when in polyinit mode
This can be classified as experimental, since I am only bothering to
implement it for the tty windowport. I imagine there are a handful of
games where someone has logged back into hardfought and started their
first game of xNetHack in a while only to realize that they had polyinit
set in the rc file, and now they have to play a game they didn't really
want to. This will at least give players a chance to realize that and
quit out before actually starting the game.