Recolor priests and high priests to yellow
This is a change introduced in EvilHack, and was made possible without
having color conflicts after I reverted the wererat being yellow change:
prior to this commit, yellow @ was unused. It's a sensible change; being
able to tell a priest apart from other white/gray @ (like, say, watchmen
in Minetown and player monsters in Astral) at a glance is useful.