Enhance feedback from detect unseen
I was never too happy with how this was a silent effect that required
you to watch the map to see if anything changed. It might count as
an accessibility issue as well, not sure.

This change adds specific feedback for all the possible things that
might get revealed by detecting unseen. If you reveal a secret door or a
trap, you now get a message indicating that.

If you identify the wand of secret door detection by engraving and do
reveal some things, the relatively new message about finding hidden bugs
on the floor comes off as a little awkward. It would be best if that
message could be suppressed when things get revealed, but whether things
were revealed doesn't get passed up through zapnodir().