Add greater variety to several Minetown layouts
Loosely based on a series of Fourk commits with the same aim - some of
the Minetowns were very similar looking and hard to tell the difference
between, making it a more boring experience to explore to and around
there. While there are a couple very distinctive minetowns like Orcish
Town and Grotto Town (neither of which got changed in this commit,
incidentally), most of the rest could stand to get spiced up a bit. See
the changelog for the specific adjustments.

Fourk tweaked things so that each Minetown has a general store for price
ID purposes. Since Orcish Town is in play here, though, that doesn't
work out. Izchak's shop is still guaranteed in all towns, and can still
be used for emergency price ID in the event that there are no other
shops in the dungeon.

Another principle embraced by Fourk but less thoroughly here is the
enlarging of 1-wide passageways. Alley Town retains them; I didn't feel
like it was important enough to pursue the same thing in Bazaar Town.

I also tried an interesting experiment with shop distribution in Bazaar
Town: I wanted to add a big variety of different shops, but didn't want
to have the town possibly generate all of the shops, or none of them,
which is what you get when the shops all have their own independent
probabilities. That was the best way to do things back before we had
variables and string-indexed tables and all the fun things provided by
Lua, but now that we do there's more room to try something new. So it
now tries to provide a consistent, moderate number of shops selected
from the larger pool, but weighted such that if you get less-useful
shops you will get more of them. Initially I used allotment = 100, but
then found that it ended up making about 4-6 shops every time which is
WAY more than generate now. With allotment = 70 it tends to generate
3-4 with the occasional 2 or 5.