Increase the size of struct rm
This has been a long time coming, as the number of interesting things
(and a couple fixes) that could be implemented if only we had more space
in this struct has continued to grow. By expanding it here, the
groundwork is laid.

When talking about plans to do this, I've heard scary stories from other
people versed in the code (even a variant dev) about making the save
files unbearably large. Some basic testing confirms that is hyperbole.
Comparison of save files containing the same amount of dungeon levels
before and after this change shows at most several KB difference, which
is roughly in line with what you would expect with ROWNO * COLNO *
number of levels. Over an entire, complete dungeon, the total size
increase would be at most 123 KB. (Note also that save files are
ultimately temporary - the real long term space consumer is ttyrecs).
Additionally, the impact on memory use is negligible since only one
level is held in memory at a time.

This does increase the number of bytes from 8 to 9, and I confirmed that
sizeof(g.level.locations[0][0]) gets padded up to 12 on my machine, but
struct rm is primarily used only in g.level.locations, and I expect
modern compilers to know how to pack an odd-sized struct into an array
with minimal wasted space.