Fix: Remove overloading of D_WARNED and D_TRAPPED flags
This is, in vanilla, the one buggy behavior that results from not a
coding error, but the constraint to keep rm.flags within 5 bits: there
was simply too much information about doors to cram into such a small
chunk of data. This led to the infamous behavior of a watchman yelling
at you for picking a lock on a door (setting its D_WARNED flag so that
they will anger if they catch you picking the same lock) and in the
process creating a trap on the door that wasn't there before. The main
reason for allocating a 6th bit here is to clear up this overload.

Granted, in xNetHack, Minetown-level door traps are less explosive than
the standard vanilla one, but that is no excuse for leaving it in.