Add a #terrain option to show spaces the hero has visited
New feature #1 from the List Of Stuff That Needed To Wait Until Struct
Rm Was Expanded. A couple variants (I think NetHack4 and its
derivatives) also track where the player has been on the map and offer
to show it to them; however, to my knowledge they display it in-game as
the player moves around the map. Which surely has some merit, but it
sets up certain conflicts that are hard to resolve (what is a
sufficiently unobtrusive color to use? how to indicate the player has
traversed over a colorful dungeon feature like a throne?) A #terrain
menu option seemed better.

This feature is not as fully polished as it could be. A "proper"
implementation would leave the underlying terrain symbols (stairs,
corridors, doors, features, etc) there for the player to see, and added
a new set of glyphs for each possible map character that #terrain can
show. Or possibly a new MG_ flag? Instead, this overwrites all the
visited spaces with a single glyph, orange magic platform, which I chose
because it's rarely used and readily visible to both ASCII and tiles
players. Expediency is the main reason for this; I like this feature but
not enough to spend large amounts of time sorting out how to fit all
those new glyphs into the display code.

Sorting out the various cases of where the visited flag should and
should not be set was rather annoying. (Easier solution would have been
just to count hurtled-over spaces as visited; but thankfully I figured
out how to omit them.)