Refactor: encode artifact material in artilist
This is the proper way to store artifact materials. Initially when there
were only a couple with non-default materials, it was okay to have the
materials hardcoded in oname() artifact creation, but as the list of
artifacts with non-standard materials has grown, it's gotten more
unwieldy. Additionally, the idea of having "artifact lookup" to go along
with monster and object lookup is attractive, and for that the material
really needs to be in the statblock.