Add artifact Apple of Discord
This is a first in a few important ways. Mainly, it is the first
non-weapon non-quest artifact in xNetHack. As such, it can be a
sacrifice gift, and I was surprised to only find one assumption in the
code that any non-quest artifact would be a weapon.

The other main first is that it's the first artifact food (not EVER
across all variants, but certainly not that many variants have
implemented artifact food. I don't even think *dNetHack* has any). This
created some interesting issues during testing, since you can of course
eat the artifact. Could have let it be inedible, or edible only to
metallivores, but I had behavior in mind for eating it, so eating it
stands.

The initial plan for its powers was to give infinite hungerless conflict
when carried (using a new SPFX_CONFLICT) to crowned chaotics, and
invokeable hungerless conflict to all others. However, I decided to
simplify the implementation by having it be invokeable hungerless
conflict for everybody, but it boosts the strength of that conflict for
crowned chaotics. (Actually it just boosts all conflict when a crowned
chaotic is carrying it; even if the Apple itself is not generating
conflict, a ring or Big Stick will have the same amplified effect). This
gives a bit more incentive to get crowned if you're chaotic.

And eating it gives you certain hungering conflict, same as eating a
ring of conflict does. This seems suboptimal enough to me that I think
players would mostly still carry it around, but hopefully it isn't a
completely foregone conclusion and there would be some people who decide
they'd rather eat it instead. (Not implemented: any guards against other
monsters eating it. Confirmed that monsters that eat apples will indeed
eat it, and metallivores seem to ignore it. Gelatinous cubes engulf
without eating. So the only risk of it getting eaten seems to be from
non-carnivorous pets that are hungry.)

In terms of game balance, this is sort of a half step towards adding
so-called "minor artifacts" that are strictly less powerful than the
most powerful ones you can currently receive from the gods. Conflict
without the main drawback of the ring is certainly good to have, but at
the same time this doesn't directly make the hero better at fighting.
Hopefully it hits my target of being a step down in power from the best
vanilla non-quest artifacts, while not being obvious about it.