Fix: whereis used pointer to static memory as truth value
Using never-zero pointer address of g.killer.name as a boolean was
incorrect but was only detected by gcc 12. Probably what was intended
was g.killer.name[0], but it doesn't seem completely correct to use that
either, since I don't think it always gets cleared after being written
to (normally it isn't relevant if the game is still going). So instead
use g.program_state.gameover, which is only set in really_done from
which there is no exit, and I doubt whereis gets updated between then
and ending the process but at the same time if you used something that
was set *before* really_done is called, there's a possibility the hero
gets lifesaved.

The "ascended" value in whereis doesn't actually seem to be used
anywhere, so this may all be pointless.