Fix: buffer overflow when entering long epitaph
Providing a char[101] buffer to getlin didn't actually limit user input
to 100 characters; instead it ensured anything longer would end up
written past the end of the buffer.  Not possible in TTY (since for some
reason getlin input length is limited to COLNO, getline.c:146) but in
curses this could be done and cause a crash on some platforms, or
potentially be put to some other malicious end.  Prevent this from
happening (and remove the 100 character limit, since it's not actually
enforced when entering text; instead, limit the full epitaph to the size
of the buffer).

I also changed things around so that selecting to enter your own
epitaph, but then entering an empty line, will use the default
formatkiller headstone engraving.