Avoid "killed by a died" from open air fall
The second call to done() if falling into a bottomless pit (used only if
the hero was lifesaved for the first one, then no safe teleport
destination could be found) could produce "killed by a died" as the
death reason.  Since g.killer.name is reset by done() after the hero
survives, and wasn't being set again before the second call, it would be
left empty and the (somewhat nonsensical) default death reason would be
used.  I don't think there are currently any bottomless pits in the
game, so this was a latent bug.