Don't grant tame djinni from lamps/potions if player is petless
This is similar to the Astral angel not appearing if the player is
petless. Strictly speaking, players attempting to play petless could
control this simply by not rubbing any lamps or quaffing any smoky
potions, but that has a rather large impact on the game just for
maintaining a conduct.

It doesn't change the overall probability of a hostile or wish-granting
djinni; a petless player will simply get the neutral "You freed me!"
djinni twice as often.

This could technically deny a djinn pet from players who actually wanted
one, but since such players would have to have the pet option turned off
and acquire no other pets along the way to making the djinn come out,
it's probably an extreme edge case.