Implement boss-blocking of magic mapping
One main motivation of this is for the Quest, preventing the hero from
peeking ahead on certain levels (such as the Archeologist locate, lower
filler, and goal levels) to see how the terrain got generated and where
the hidden secrets are. But at the same time, it doesn't really make
sense to mark the levels as strictly unmappable, because the nemesis may
be the force preventing you from mapping them. And it may not work just
to clear the nommap flag, since you could kill the enemy off-level.

The solution here is to expand the nommap flag to 2 bits to allow a
third value, "boss blocked", which if magic mapping is attempted checks
whether the local boss has been defeated and if not will not allow you
to map. It works somewhat similarly to demon lord teleport blocking but
not quite the same: demon lords take their teleport blocking to any
level they are on, whereas here it only applies in the lair but still
applies if the enemy is alive but off-level. (Otherwise it could be
pretty easy just to quaff cursed invisibility, let the enemy warp to
you, take them upstairs and return without them, and then map the level
in their absence.)

Note that it's up to the level designer to apply this new flag; if left
unspecified the level will still be mappable even if there is a boss
present, and if specified as "nommap" it will still be never mappable
even after the boss has been killed.

I have applied this change to the four demon lord lairs and certain
Quest levels I have redesigned. Quest levels I have not touched remain
the same as in vanilla. This could be extended for other boss lairs like
the Wizard's Tower, Medusa, and Vlad's Tower, but none of those felt
particularly important to prevent mapping on.