Make gold items a loadstone-esque curse magnet
This is intended to increase the differentiation between object
materials by adding another unique dimension to objects made of gold. In
folklore and myth, gold treasures are likely to be cursed. So, the
beneficial part of this change is that gold accepts curses far more
easily than anything else to the extent that it will take curse item
effects before any other items in the player's inventory. This allows
players to deliberately keep gold items in their inventory to protect
other items from getting cursed.

The drawback to this is that the curse on gold items prevents you from
getting rid of it. This piggybacks on the previously empty undroppable()
macro left behind when loadstones were removed, so cursed gold items act
pretty much exactly like loadstones. They can be stolen but not
voluntarily dropped or thrown away, and even the weakest remove curse
effects will lift the curse from all gold items in the inventory.