Allow entry to the Quest at level 1 if you have never gained a level
This is from Fourk, with minor code modifications and less minor quest
text modifications. The main reason for doing this is to make things
easier on pacifist players who would otherwise have to go through a lot
of tedium to farm wraith corpses or potions of gain level and probably
wouldn't have the ability to enter the Quest until after Gehennom.
It's a particular problem in xNetHack because getting levels off a
foocubus is far less reliable and farmable than it is in vanilla.

Note that this specifically requires the level max to have never
advanced beyond 2; if you're rejected for inexperience at levels 2-9,
you can't drain yourself to 1 and get accepted.