Make spheres' explosions passive as well as active
Completing the other half of the "spheres make real explosions"
addition, and the final third of the "buff spheres because they're
pretty useless" set of changes. Unlike making their active attack a real
explosion which involved a lot of refactoring, this change is as simple as
giving them an AT_BOOM attack.

Reason for this is that it doesn't make a ton of sense that spheres are
these weird organisms whose entire biology is to blow up and end its
life, but if you hit it particularly hard with a sword and killed it
first, it would fail to blow up.

To compensate for the added wrinkle of this change, flaming spheres and
shocking spheres are ticked up to difficulty 10 and 11, respectively.
Freezing spheres remain at 9.