Dwarf and gnome monarchs ignore worthless glass
I tried to figure out a way to do this without refactoring the mitem_*
functions to take the monster as well as the item, but everything I came
up with ended up more convoluted or less futureproof than that, so
that's what will stick.

Players in the know will be able to use this to informally
differentiate valuable and worthless glass. I plan to add a couple
true rumors indicating this behavior as well.