Inform player and abuse Wisdom when throwing weapon incorrectly
If someone has paranoid throw off, they can end up inexpertly
throwing their stack of arrows by hand, or whatever. This increases the
feedback for that by adding a "You clumsily throw..." message. It also
abuses Wisdom some of the time. Granted, everyone throws the occasional
sword at a floating eye, but habits shouldn't be made of it.

The other change in here is the expansion of this system beyond ammo to
any weapon deemed unsuitable for throwing (see the comments/changelog for
more info). Before anyone complains about tridents not being counted as
a throwing weapon here, objects.h specifically says they're "not
intended to be thrown".