Intelligent monsters pick up keys and lock picks
This is from UnNetHack and allows a way for an arbitrary monster to
acquire an unlocking tool to start opening doors (and containers, if
vanilla eventually implements this) with.

One note is that gnomes in the Mines may pick up ones that randomly
generated, making them somewhat rarer for players; but this only seems
to happen when the gnomes are hostile, and then the player is likely to
kill them anyway.