Implement behavior for cursed potion of gain ability
This potion doing nothing whatsoever if you drink it when cursed is
pretty weak. It ought to have some effect, even if the effect isn't
particularly great in keeping with how cursed effects usually are.

This implements an idea by FIQ that the potion should gain you an
ability point by stealing the point from elsewhere. The trick is that it
will always grant to the lowest score, and steal from the highest -
which could make it situationally useful for raising low ability scores
when your highest score is not of particular importance (think very high
Charisma or 18/xx Strength) and you lack the resources to uncurse or
bless the potion.