Gehennom overhaul: restructure the dungeon
This first step puts all the special levels and filler templates in
place, and sets up all the new branches to connect together properly.
I'm not 100% on all the branch names, but most of them seem good enough,
and I avoided using "Sheol" for any of them to avoid confusion with
UnNetHack.

As of this commit, levels most certainly do NOT work correctly -- the
goal for it was to get it into a state where the various new branches
are traversible, without erroring out on dungeon misconfiguration or lua
files not being found. Thus, this takes care of everything related to
configuring dungeon.lua and adding new lua files for level templates.
But all the new levels are currently empty bigrm-1 clones, and the
Wizard's Tower is broken because it still assumes it's embedded in a
filler maze in the main branch rather than its own branch.

Since I haven't come across a case yet where someone has actually needed
the "Files" file for something (more or less anyone who wants to get the
xNetHack source has enough space to download even the system files
they're not using), I added a note to Files stating that it's not
maintained.