Overhaul the Wizard's Tower
This structure has been in the game for a while, but hasn't changed much
since it was changed from the original mini-tower design. To my
knowledge, no variant has really attempted to change it thus far,
so it tends towards the same experience even across different variants.
More to the point, its fixed layout means that it doesn't really present
much of a challenge to anyone who has played through it a few times.
Complicating that matter is that it's embedded into a larger level,
which is neat, but constrains the available space with which to work.
(Also, the embedding as it existed necessitated a 1-wide maze, which is
being eradicated almost everywhere in xNetHack Gehennom).

Earlier commits already moved the Wizard's Tower to its own branch
(which incidentally technically puts it "outside Gehennom" - for the
moment, I'm going to leave that as it is). This throws out the levels
entirely, replacing their filler maze and small contained tower with a
large, circular tower level. Similar to other branches I have designed,
the tower contains cues to give it a more realistic climbing-a-tower
feel, such as corresponding ladders being in the same spot on the map.

The first floor is a new sort of puzzle that involves triggering a
mechanism that opens and closes paths into the center of the level
(where the ladder to the next level is). I had several false starts and
ideas that went nowhere in the process of designing this puzzle, but
think it's in a pretty good shape now. It manages to be an interesting
thing for players to figure out without adding new terrain or objects or
anything. Still, as of this commit, it may still be a little too easy to
brute force, and I might yet change it.

The second floor is a gauntlet run of monsters - the treasure zoo and
beehive that previously existed in the Wizard's Tower, plus some other
groups of monsters that it seems like the Wizard would keep around as
flunkies.

The final floor contains the Wizard himself, at the end of a maze of
rooms connected with secret doors and containing a number of both
goodies and baddies.

Saves are broken again in this commit due to adding data to track the
new puzzle to the save file.

Other changes in this commit:
- Got rid of a bunch of code that assumed the Wizard's Tower was
  embedded inside Gehennom. In_W_tower() is gone, but there is still
  On_W_tower_level().
- Externified trapnote().