Retool the Wizard's Tower puzzle for multiple rings (2)
As mentioned in the previous commit message, having a single ring meant
that it was pretty darn easy to "solve" the puzzle just by stepping on
every board without making any effort to figure out the mechanism, since
any given trigger had a roughly 1/8 chance of solving it by random luck.
I tested a bunch of times going around the circle stepping on every
board, and more often than not I had "solved" the puzzle before making
one complete circle.

The attempted fix in this commit is to add a second ring, which will
reduce the chance of randomly solving the puzzle to 1/64. (A lot of the
functionality in this is designed with an arbitrary number of rings that
happens to be set to 2, but that is less because I have any plans of
adding a third ring and more to avoid hardcoding.)

Every switch now triggers the gap in both rings to move. The puzzle is
solved when you step on a switch that moves both rings to the chamber
you are in. I've played through it and have confirmed there is enough
information given in the clues to work out a solution; but I think this
formulation of the puzzle may prove too complex for players' enjoyment.
A potential less-complex version of the puzzle would be to have every
switch leave one ring alone and rotate the other ring 1 or 2 rooms.

One other adjustment in this commit is to disallow monsters stepping on
the squeaky boards from triggering the puzzle, since this is just
incredibly annoying in practice when you're trying to hit specific
boards in a specific order.