Implement specific-destination teleport traps
For an upcoming planned feature for the Wizard's Tower. These are
regular teleport traps, but they will always teleport the player or a
monster who jumps into them to the same square, every time, rather than
being random. (Not really like vault teleporters, which search for a
vault when you activate them, and then disappear.)

The only way for these to generate is when explicitly specified using
"teledest={coord}" in the table form of des.trap() in a special level
file. I have no plans to allow them to randomly generate.