Turn the Wizard's Tower second floor into a teleportation maze
This is because "run a gauntlet of various predefined monster types" is
boring and probably several of the other gehennom levels will end up
with elements of that, whereas none of them are probably going to end up
with anything like this. The previous commit laid the groundwork with
its fixed-destination teleporters; the main effort here was creating a
Lua algorithm for generating the maze.

I've tested the layout of the maze to be confident that it generates the
properties I wanted (single solution path with no shortcuts, no
unreachable rooms, and every room has a path back to start) but am not
sure those parameters will generate a level that is actually fun to
solve. It may require reducing the amount of teleporters or rooms.

Other changes in this commit add a macro Is_telemaze_level() and special
case teleport traps on this level so that the hero is not prompted
before stepping into them. Also, to make the teleport traps work, it
obviously isn't possible to have the noteleport flag on this level, but
to prevent the player from just controlled-teleporting to the ladder up
to bypass it, teleport control no longer works within the Wizard's
Tower. Random teleports do still work, so with enough attempts, one
could eventually hit the (rather small) goal room.