Fix: monsters given weapons in Lua did not actually use them
Noticed with a demon lord, but I think it generalized to any monster who
gets their weapon given to them as part of their inventory in a special
level file. This appears to be because their weapon_check is never
updated, so they fail to consider that they have a better option than
punching opponents barehanded.

In xNetHack (but not in vanilla) monsters that init with a weapon in
makemon() automatically wield it; this extends that to monsters that
init with a custom inventory. I ended up refactoring that system; now it
uses a new codepath for NEED_WEAPON in mon_wield_item which takes care
of determining whether the monster should wield a melee or ranged
weapon.

In the process, fixed a couple of additional bugs:
- Monsters that initted with a ranged weapon in makemon would not use
  it (at least until the hero came within range and their AI forced them
  to consider a ranged weapon), because the logic I added in commit
  67d9fcd hardcoded NEED_HTH_WEAPON. Now it follows the agnostic
  NEED_WEAPON logic which always selects a ranged weapon when that's the
  only option.
- Monsters immediately wielding a weapon produced the usual wield
  messages, even if the monster was being created at level creation
  time, so one could generate a new level and immediately get e.g. "The
  goblin wields an orcish dagger!" Now, the messages are suppressed
  during level creation.