Implement the City and Citadel of Dis
As planned, this is taking over from Orcus as the second-half inhabited
area in the main branch of Gehennom. Challenging Dispater remains
optional; you have to enter the Citadel and travel up one floor if you
want to reach him. But most players will still have to traverse the
city.

One theme that does not carry over from Orcus's level is the undead;
while there is a graveyard here and there are some roaming ghosts and
shades, it's not a city /of/ the undead. Orcus's level will keep that
theme. Instead, Dis is the city of demons, and is accordingly mainly
populated by them.

The three erinyes stand guard outside the gate, and they can no longer
generate randomly or in graveyards.

I added a seminary of Moloch (rather than an abandoned temple) at the
last minute, but implemented in Lua rather than C, so it's a temple as
far as game internals are concerned. The Gehennom special rooms are
nice, but they're probably going away since they were designed for
filler levels.