Don't generate bribing monsters as invisible
This particular behavior never made sense to me. The invisibility is
supposed to be stripped as soon as the monster appears to make its
bribe, but this doesn't really seem to work in xNetHack, especially
since some bribing demons are not going to be covetous warpers anymore.

Additionally, I think it interfered with some xNetHack demon entrance
messages, where the message wouldn't be delivered if the player can't
see who's delivering it.