Overhaul Dispater's behavior
Even with the level redesigns, actually _fighting_ him was not any
different from normal -- he would warp to the stairs once he got banged
up a little bit, the player would chase him there, and then they'd trade
melee blows until he died. Not interesting.

He now pursues a strategy of defense and combat avoidance, based on his
portrayal in the Forgotten Realms D&D setting. To this end, I removed
his covetous warping (it would be nice if a monster could be covetous
and still pursue coveted items but not warp to them, but this is not
worth implementing right now) and implemented a special spell system
that gives him a custom spell list, with two new spells that he can cast
to get away from the hero and slow them down. Like Schliemann, he also
tries to keep away from the hero when they close to melee combat (though
he doesn't try to stay out of a direct line with the hero like a
unicorn will).

Since he is fairly easy to access from the main path through Gehennom,
the intent is for this to be fairly difficult for an optional boss
fight.

The current plan is for him to remain a demon lord that will accept
bribes, but I plan to update the bribing code to make it a tougher
choice for the hero.