Remove unnecessary exclude_islev=1 from various level files
I think this was a form of cargo culting I developed after probably
running into the issue with placing a levregion at 0,0 on the map that
the previous commit fixes, finding that it went away if exclude_islev=1
was specified, and copying that into subsequent regions.

Now that leaving exclude unspecified forces exclude_islev=1 and the
exclusion region off the map, these should no longer be needed.