Implement the Gate of Hell level
This is a main-branch Gehennom level, and its position directly below
the Valley means that it is the "proper" entrance to Gehennom, whereas
the Valley is more like the foyer. (And descending the stairs from the
Valley means the "unspeakable cruelty and harm" message refers to this
level and everything below, which works well.)

What I'm going for here is the feeling that you're now actually kicking
in Moloch's front door, a sense which isn't really present in vanilla.
Welcome to the next phase of the game.

This level features Cerberus, of course, and also is the branch level
for Vlad's Tower, the stairs to which are off to one side. Other than
that, pretty standard Gehennom fare: lava, demons, a couple dragons.
This doesn't need as strong of a theme, I think, as the demon lord lairs
need to have.