Gate of Hell drawbridge does not open while Cerberus is guarding it
Unlike other drawbridges, this one will not open or be destroyed until
Cerberus is dealt with. Killing him is sufficient, but not necessary -
he can also be subdued by e.g. putting him to sleep.

One reservation I have about this setup is that some players might see
Cerberus (or even not visit the island he's on and not notice him at
all) and get to the drawbridge then get confused when they can't open
it. I did add a true rumor noting this behavior in this commit, and am
hoping that the statues of hell hounds that flank the gates should be a
clue as well. Originally, Cerberus was in the area right in front of the
gate and impossible to miss, but I realized this made it pretty trivial
to float over the lava out of his reach and bounce rays off the wall
behind him to take him down.