Add a lua command to clear a map/room's relative coordinate system
The next special level I am planning involves a swamp-style level_init,
followed by small, randomly placed maps, followed by some randomly
placed monsters/traps/objects which can go anywhere on the level. But
the current state of things did not allow this to work:
- If I place the monsters, traps and objects before the maps, the maps
  may overlap them and leave monsters embedded in a wall or traps on
  invalid pieces of dungeon furniture, etc.
- If I place the maps first, des.[monster|trap|object|...] will place
  them randomly within the last defined map. They won't appear randomly
  on the level.

So what was needed was a way for the level designer to tell the parser
to abandon the coordinate system of the last defined des.room or
des.map, which this implements.

A more proper implementation might be to insist that things intended to
be placed relative to a room or map should always go inside the contents
function of that room or map, and reset the coordinate system to
level-wide whenever not currently inside either, but that would break a
ton of levels which assume the current behavior of everything after a
map being relative to it even though it's not part of contents.

This commit is temporary! It will be superseded by 4af086b in vanilla
which added reset_xystart_size() automatically to the end of des.room()
and content-function-less des.map().